# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baskonia Pipeline API",
    "description": "API de análisis de Baskonia (fase F3 de la migración strangler-fig).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/teams/{slug}": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Get Team Detail",
        "description": "Detalle de un equipo por slug (endpoint 4).",
        "operationId": "get_team_detail_api_v1_teams__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{slug}/filters": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Get Filters",
        "description": "Filtros disponibles para un equipo: temporadas y competiciones (endpoint 5).",
        "operationId": "get_filters_api_v1_teams__slug__filters_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiltersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{slug}/summary": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Get Summary",
        "description": "Resumen del equipo: identidad, filtros y medias avanzadas (endpoint 6).",
        "operationId": "get_summary_api_v1_teams__slug__summary_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Año de inicio de temporada (p.ej. 2025)",
              "title": "Season"
            },
            "description": "Año de inicio de temporada (p.ej. 2025)"
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Competición: acb | euroleague | supercopa",
              "title": "League"
            },
            "description": "Competición: acb | euroleague | supercopa"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvancedSummary": {
        "properties": {
          "avg_pace": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Pace"
          },
          "avg_off_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Off Rating"
          },
          "avg_def_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Def Rating"
          },
          "avg_net_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Net Rating"
          },
          "avg_efg_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Efg Pct"
          },
          "avg_ts_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Ts Pct"
          }
        },
        "type": "object",
        "title": "AdvancedSummary",
        "description": "Medias de estadísticas avanzadas de un equipo (cada clave puede ser null)."
      },
...
      "FiltersResponse": {
        "properties": {
          "seasons": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Seasons"
          },
          "default_season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Season"
          },
          "leagues": {
            "items": {
              "$ref": "#/components/schemas/LeagueOption"
            },
            "type": "array",
            "title": "Leagues"
          }
        },
        "type": "object",
        "required": [
          "seasons",
          "default_season",
          "leagues"
        ],
        "title": "FiltersResponse",
        "description": "Filtros disponibles para un equipo (cabecera de la app)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeagueOption": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "code",
          "label"
        ],
        "title": "LeagueOption",
        "description": "Opción de competición para el selector de filtros."
      },
...
      "SummaryResponse": {
        "properties": {
          "team": {
            "$ref": "#/components/schemas/TeamRef"
          },
          "filters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filters"
          },
          "advanced": {
            "$ref": "#/components/schemas/AdvancedSummary"
          },
          "games_played": {
            "type": "integer",
            "title": "Games Played"
          },
          "games_upcoming": {
            "type": "integer",
            "title": "Games Upcoming"
          }
        },
        "type": "object",
        "required": [
          "team",
          "filters",
          "advanced",
          "games_played",
          "games_upcoming"
        ],
        "title": "SummaryResponse",
        "description": "Resumen del equipo: identidad, filtros aplicados y medias avanzadas."
      },
      "TeamRef": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name"
        ],
        "title": "TeamRef",
        "description": "Referencia a un equipo (slug + nombre para mostrar)."
      },
      "TeamResponse": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "league": {
            "type": "string",
            "title": "League"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "league"
        ],
        "title": "TeamResponse",
        "description": "Equipo con su liga de origen."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}